+++ src/main/java/de/alpharogroup/spring/batch/mapper/CustomBeanWrapperFieldSetMapper.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package de.alpharogroup.spring.batch.mapper;
 
 import java.beans.PropertyEditorSupport;
@@ -14,77 +38,108 @@ import lombok.NonNull;
 import lombok.experimental.FieldDefaults;
 
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public class CustomBeanWrapperFieldSetMapper<T> extends BeanWrapperFieldSetMapper<T> {
+public class CustomBeanWrapperFieldSetMapper<T> extends BeanWrapperFieldSetMapper<T>
+{
 
-    @NonNull
-    DateTimeFormatter formatter;
+	@NonNull
+	DateTimeFormatter formatter;
 
-    public CustomBeanWrapperFieldSetMapper(@NonNull Class<? extends T> type, @NonNull DateTimeFormatter formatter) {
-        setTargetType(type);
-        this.formatter = formatter;
-    }
+	public CustomBeanWrapperFieldSetMapper(@NonNull Class<? extends T> type,
+		@NonNull DateTimeFormatter formatter)
+	{
+		setTargetType(type);
+		this.formatter = formatter;
+	}
 
-    @Override
-    protected void initBinder(DataBinder binder) {
-         binder.registerCustomEditor(LocalDateTime.class, new PropertyEditorSupport() {
-            @Override
-            public void setAsText(String text) {
-                if (!text.isEmpty()) {
-                    setValue(LocalDateTime.parse(text, formatter));
-                } else {
-                    setValue(null);
-                }
-            }
+	@Override
+	protected void initBinder(DataBinder binder)
+	{
+		binder.registerCustomEditor(LocalDateTime.class, new PropertyEditorSupport()
+		{
+			@Override
+			public String getAsText()
+			{
+				Object date = getValue();
+				if (date != null)
+				{
+					return formatter.format((LocalDateTime)getValue());
+				}
+				else
+				{
+					return "";
+				}
+			}
 
-            @Override
-            public String getAsText() {
-                Object date = getValue();
-                if (date != null) {
-                    return formatter.format((LocalDateTime) getValue());
-                } else {
-                    return "";
-                }
-            }
-        });
-        binder.registerCustomEditor(LocalDate.class, new PropertyEditorSupport() {
-            @Override
-            public void setAsText(String text) {
-                if (!text.isEmpty()) {
-                    setValue(LocalDate.parse(text, formatter));
-                } else {
-                    setValue(null);
-                }
-            }
+			@Override
+			public void setAsText(String text)
+			{
+				if (!text.isEmpty())
+				{
+					setValue(LocalDateTime.parse(text, formatter));
+				}
+				else
+				{
+					setValue(null);
+				}
+			}
+		});
+		binder.registerCustomEditor(LocalDate.class, new PropertyEditorSupport()
+		{
+			@Override
+			public String getAsText()
+			{
+				Object date = getValue();
+				if (date != null)
+				{
+					return formatter.format((LocalDate)getValue());
+				}
+				else
+				{
+					return "";
+				}
+			}
 
-            @Override
-            public String getAsText() {
-                Object date = getValue();
-                if (date != null) {
-                    return formatter.format((LocalDate) getValue());
-                } else {
-                    return "";
-                }
-            }
-        });
-        binder.registerCustomEditor(LocalTime.class, new PropertyEditorSupport() {
-            @Override
-            public void setAsText(String text) {
-                if (!text.isEmpty()) {
-                    setValue(LocalTime.parse(text, formatter));
-                } else {
-                    setValue(null);
-                }
-            }
+			@Override
+			public void setAsText(String text)
+			{
+				if (!text.isEmpty())
+				{
+					setValue(LocalDate.parse(text, formatter));
+				}
+				else
+				{
+					setValue(null);
+				}
+			}
+		});
+		binder.registerCustomEditor(LocalTime.class, new PropertyEditorSupport()
+		{
+			@Override
+			public String getAsText()
+			{
+				Object date = getValue();
+				if (date != null)
+				{
+					return formatter.format((LocalTime)getValue());
+				}
+				else
+				{
+					return "";
+				}
+			}
 
-            @Override
-            public String getAsText() {
-                Object date = getValue();
-                if (date != null) {
-                    return formatter.format((LocalTime) getValue());
-                } else {
-                    return "";
-                }
-            }
-        });
-    }
+			@Override
+			public void setAsText(String text)
+			{
+				if (!text.isEmpty())
+				{
+					setValue(LocalTime.parse(text, formatter));
+				}
+				else
+				{
+					setValue(null);
+				}
+			}
+		});
+	}
 }