+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Change log
 ----------------------
 
+Version 1.1
+-------------
+
+ADDED:
+
+- new inteface for map entities to view models and back
+- new spring abstract rest controller created
+- new service interface created that delegates all calls to the JpaRepository
+
 Version 1
 -------------